Refactor flow container to just connect and assign props.
