+++ packages/venia-concept/src/components/Checkout/__tests__/container.spec.js
@@ -1,6 +1,3 @@
-import React from 'react';
-import testRenderer from 'react-test-renderer';
-
 import {
     beginCheckout,
     cancelCheckout,
@@ -10,9 +7,8 @@ import {
     submitPaymentMethodAndBillingAddress,
     submitShippingMethod
 } from 'src/actions/checkout';
+
 import ConnectedCheckoutContainer from '../index';
-import { CheckoutContainer } from '../container';
-import Flow from '../flow';
 
 jest.mock('src/classify');
 jest.mock('src/drivers', () => ({
@@ -31,20 +27,6 @@ jest.mock('src/drivers', () => ({
         return component;
     }
 }));
-
-const defaultProps = {
-    checkout: {
-        step: 'cart',
-        submitting: false
-    },
-    beginCheckout,
-    cancelCheckout,
-    editOrder,
-    submitShippingAddress,
-    submitOrder,
-    submitPaymentMethodAndBillingAddress,
-    submitShippingMethod
-};
 
 test('returns a connected CheckoutContainer component', () => {
     expect(ConnectedCheckoutContainer.component).toBeInstanceOf(Function);
@@ -77,32 +59,4 @@ test('mapStateToProps correctly maps state to props', () => {
         checkout: state.checkout,
         directory: state.directory
     });
-});
-
-test('returns a Flow component', () => {
-    const props = {
-        ...defaultProps,
-        cart: {
-            details: {}
-        }
-    };
-    const component = testRenderer.create(<CheckoutContainer {...props} />);
-
-    expect(() => component.root.findByType(Flow)).not.toThrow();
-});
-
-//TODO: I'm conflicted. The propTypes indicate that cart and checkout should
-// always be defined as objects however the check in the wrapper component is
-// for existence. If either of these props are falsy then we'd get a propType
-// warning. So, should we remove this conditional check and rely on propTypes
-// or should I just write a test and be OK with it warning every test run?
-test('does not render Flow component if cart/checkout are falsy', () => {
-    const props = {
-        ...defaultProps,
-        cart: false,
-        checkout: false
-    };
-    const component = testRenderer.create(<CheckoutContainer {...props} />);
-
-    expect(() => component.root.findByType(Flow)).toThrow();
 });
